fix: rollback Decimal-ization, fix compile errors and high-severity issues
Reverts the rust_decimal migration that broke compilation (sqlx SQLite
deliberately omits Decimal support). Restores f64 for all monetary fields
and fixes the bugs and security issues introduced or left open by the
previous commit.

Compile fixes (26 errors):
- Roll back OrderRequest/Order/CreateOrder/UpdateOrderFilled to f64
- Remove sqlx rust_decimal feature; rust_decimal serde feature
- Hyperliquid round_price/round_size accept f64 again (internal Decimal kept)
- Remove dead LoginSuccess struct
- Fix clippy redundant_closure warnings (auth middleware + bot dispatch)

Agent-introduced bug fixes:
- App.tsx: replace getCookie with /api/orders probe (HttpOnly cookie unreadable by JS)
- LoginPage.tsx: fix variable shadowing (const data) and code scope in SSE callback
- auth/handlers.rs: drop Secure cookie attribute for local HTTP dev compatibility

Correctness fixes:
- B1: update_fill_progress uses MAX(filled_quantity, ?) to prevent regression under concurrent fill events
- B2: add recover_pending_orders() on startup — orphaned pending orders (crash during placement) now marked failed
- B3: add cancel_order_db() with conditional WHERE status IN ('open','partially_filled') — prevents clobbering filled status; allow cancel of partially_filled orders
- B8: engine uses adapter.kind().effective_leverage() for reverse legs instead of raw cast
- B5: add is_finite() checks in API and bot to reject NaN/Infinity inputs

Security fixes:
- S2: panic in release builds if JWT_SECRET is the insecure default
- S5: downgrade login code log from INFO to DEBUG
- S7: custom Debug impl for Config redacts all secret fields
- B7: pin hypersdk git dependency to rev a55f068e for reproducible builds

Migrations:
- Consolidate 001/002/003 into single 001_init.sql (all were destructive DROP+CREATE)

+++ src/exchange/mod.rs
@@ -1,5 +1,4 @@
 use async_trait::async_trait;
-use rust_decimal::Decimal;
 use std::collections::HashMap;
 use std::sync::Arc;
 
@@ -17,8 +16,8 @@ pub mod mexc;
 pub struct OrderRequest {
     pub symbol: String,
     pub side: String,       // "buy" | "sell"
-    pub quantity: Decimal,
-    pub price: Decimal,
+    pub quantity: f64,
+    pub price: f64,
     /// Leverage. 1 for spot; user-specified (>=1) for perp. Inherited by reverses.
     pub leverage: u32,
 }
